+++ R/plotLoadings2D.R
@@ -59,6 +59,10 @@
 #' assign new colors for the number of levels it computed (with a message).  To get
 #' exactly what you want, specify both levels and colors in equal numbers.  Function
 #' \code{\link{inspectLvls}} can help you choose appropriate levels.
+#' If you specify more than one spectrum to plot, e.g. \code{which = c(1,2)}, then
+#' arguments \code{lvls} and \code{cols} must be lists of levels and colors, one list
+#' element for each spectrum to be plotted (if specified at all).  Two convenience functions exist to
+#' make this process easier: \code{\link{LofL}} and \code{\link{LofC}}. See the examples.
 #'
 #' @author Bryan A. Hanson, DePauw University.
 #'